+++ simplegen-aio/src/test/kotlin/com/ancientlightstudios/simplegen/YamlIntegrationTest.kt
@@ -2,6 +2,7 @@ package com.ancientlightstudios.simplegen
 
 import io.kotest.core.spec.style.BehaviorSpec
 import io.kotest.matchers.shouldBe
+import io.kotest.matchers.string.shouldContain
 
 class YamlIntegrationTest : BehaviorSpec({
 
@@ -18,7 +19,21 @@ class YamlIntegrationTest : BehaviorSpec({
                 outputFile.exists() shouldBe true
             }
             Then("the output file contains the correct data") {
-                outputFile.readText() shouldBe "works=true"
+                outputFile.readText() shouldContain "works=true"
+            }
+
+            listOf("test-lorem", "test-ipsum", "test-dolor" ).forEach { name ->
+                val otherOutputFile = fixture.resolveOutputFile("$name.txt")
+
+                Then("an output file is created for $name") {
+                    otherOutputFile.exists() shouldBe true
+                }
+
+                Then("the output file contains the correct data for $name") {
+                    val text = otherOutputFile.readText()
+                    text shouldContain "works=true"
+                    text shouldContain "node=$name"
+                }
             }
         }
     }