+++ simplegen/src/main/kotlin/com/ancientlightstudios/simplegen/configuration/Transformation.kt
@@ -3,12 +3,14 @@ package com.ancientlightstudios.simplegen.configuration
 class Transformation(
     val template: String = "",
     val data: Any = emptyList<Any>(),
-    val nodes: String = "",
+    val nodes: Any = "",
     val outputPath: String = "",
     val templateEngine: TemplateEngineConfiguration? = null
 ) {
 
     val parsedData: List<DataSpec> by lazy { parseData() }
+
+    val parsedNodesExpression: NodesSpec by lazy { parseNodes() }
 
     private fun parseData(): List<DataSpec> {
         val list = mutableListOf<DataSpec>()
@@ -36,6 +38,20 @@ class Transformation(
         return list
     }
 
+    private fun parseNodes(): NodesSpec {
+        // if nodes is just a string, we assume it to be a jsonpath expression,
+        // if it is a map, we extract the expression and the type from the map
+        return when (nodes) {
+            is String -> NodesSpec(nodes as String, "jsonpath")
+            is Map<*, *> -> {
+                val expression = nodes["expression"] as? String ?: ""
+                val type = nodes["type"] as? String ?: "jsonpath"
+                NodesSpec(expression, type)
+            }
+            else -> NodesSpec("", "")
+        }
+    }
+
     class DataSpec(
         val basePath: String,
         val includes: List<String>,
@@ -45,5 +61,10 @@ class Transformation(
         val file: String? = null,
         val resultPath: String = "",
         val parserSettings: Map<String, Any> = emptyMap(),
+    )
+
+    class NodesSpec(
+        val expression:String,
+        val type: String
     )
 }